Update 230801 v2
Updated migrations
Updated vue components. They're now a bit cleaner but it's a clear mindfuck.
Updated Cart and CartItems model functions.
Updated DeleteOldCarts job so it uses the model function to delete carts.
Updated the cart controller so we validate the product_id on removeItem before we actually remove it.

+++ ecom/database/migrations/2023_05_06_104747_create_cart_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->bigInteger('cart_data_id')->constrained('cart_data', 'id')->onDelete("cascade");
             $table->bigInteger('product_id')->constrained('product', 'id')->onDelete("cascade");
-            $table->integer('qty');
+            $table->bigInteger('qty');
             $table->json('product_data')->nullable();
             $table->timestamps();
         });
